scripts/requirements.txt
psycopg2-binary==2.9.5
psycopg2-pool==1.1
